move: CocktailDetail.js Component to pages
