+++ client/src/Styles/CocktailDetail.scss
@@ -1,17 +1,24 @@
 @import "~bootstrap/scss/bootstrap";
 .CocktailDetail {
-  background-color: aquamarine;
-  height: 300px;
-  .SlideDiv {
-    background-color: aquamarine;
-    height: 300px;
-    width: 400px;
+  height: 100vh;
+  .Slider {
+    height: 100%;
+    width: 100%;
+    .slick-list {
+      background-color: rgb(0, 0, 0, 0.3);
+      height: 100%;
+      width: 100%;
+    }
+    button {
+      width: 100px;
+      height: 100px;
+      zoom: 1.5;
+    }
+    .slick-prev {
+      z-index: 9999;
+    }
   }
-  button {
-    width: 100px;
-    height: 100px;
-  }
-  .slick-prev {
-    z-index: 9999;
-  }
+  //   .slick-next {
+  //     z-index: 9999;
+  //   }
 }
